add controllers for rubric, criteria and level

+++ backend/evaluation-service/src/main/java/unicauca/edu/co/sga/evaluation_service/infrastructure/persistence/entities/CriteriaEntity.java
@@ -29,7 +29,7 @@ public class CriteriaEntity {
     private String crfDescripcion;
 
     @Column(nullable = false)
-    private Float crfPorcentaje;
+    private Integer crfPorcentaje;
 
     @Column
     private Float crfNota;
